#70: Refactor VolunteerHeader into folder structure with extracted components
