+++ src/components/Dashboard/Profile/sections/VolunteerHeader/index.ts
@@ -1,0 +1,1 @@
+export { VolunteerHeader } from "./VolunteerHeader";
